refactor: remove vercel.json
